samples: updates v1beta3 samples to v1
