+++ samples/snippets/src/test/java/documentai/v1/ProcessDocumentTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package documentai.v1beta3;
+package documentai.v1;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertNotNull;
@@ -28,7 +28,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-public class QuickStartTest {
+public class ProcessDocumentTest {
   private static final String PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
   private static final String PROCESSOR_ID = "88541adc6eeec481";
   private static final String FILE_PATH = "resources/invoice.pdf";
@@ -58,13 +58,14 @@ public class QuickStartTest {
   }
 
   @Test
-  public void testQuickStart()
+  public void testProcessDocument()
       throws InterruptedException, ExecutionException, IOException, TimeoutException {
     // parse the GCS invoice as a form.
-    QuickStart.quickStart(PROJECT_ID, "us", PROCESSOR_ID, FILE_PATH);
+    ProcessDocumentBeta.processDocument(PROJECT_ID, "us", PROCESSOR_ID, FILE_PATH);
     String got = bout.toString();
 
     assertThat(got).contains("Paragraph text:");
+    assertThat(got).contains("Extracted");
   }
 
   @After
