+++ samples/snippets/src/test/java/documentai/v1/QuickStartTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package documentai.v1beta3;
+package documentai.v1;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertNotNull;
@@ -28,7 +28,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-public class ProcessDocumentTest {
+public class QuickStartTest {
   private static final String PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
   private static final String PROCESSOR_ID = "88541adc6eeec481";
   private static final String FILE_PATH = "resources/invoice.pdf";
@@ -58,14 +58,13 @@ public class ProcessDocumentTest {
   }
 
   @Test
-  public void testProcessDocument()
+  public void testQuickStart()
       throws InterruptedException, ExecutionException, IOException, TimeoutException {
     // parse the GCS invoice as a form.
-    ProcessDocumentBeta.processDocument(PROJECT_ID, "us", PROCESSOR_ID, FILE_PATH);
+    QuickStart.quickStart(PROJECT_ID, "us", PROCESSOR_ID, FILE_PATH);
     String got = bout.toString();
 
     assertThat(got).contains("Paragraph text:");
-    assertThat(got).contains("Extracted");
   }
 
   @After
